+++ header/dev/dev.h
@@ -1,0 +1,4 @@
+#pragma once
+#include "expander.h"
+
+LRESULT CALLBACK ChildWindowProc(HWND hWnd, uint msg, WPARAM wParam, LPARAM lParam);
